+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
   * ResponseBody request filter and assertion
+  * Regex support for expected value of assert.attribute
   * Major refactor in request filter and request assertions
 
 0.7.6 / 2018-06-13